+++ SamplesApp/SamplesApp/Features/DataChannel/DataChannelConfigViewController.swift
@@ -20,6 +20,9 @@ class DataChannelConfigViewController: UITableViewController {
   /// 接続時のカメラ有効設定を切り替えるためのコントロールです。
   @IBOutlet var cameraEnabledOnConnectSegmentedControl: UISegmentedControl!
 
+  /// 開始時のマイク有効設定を切り替えるためのコントロールです。
+  @IBOutlet var microphoneEnabledOnConnectSegmentedControl: UISegmentedControl!
+
   /// 動画のコーデックを指定するためのコントロールです。
   @IBOutlet var videoCodecSegmentedControl: UISegmentedControl!
 
@@ -194,6 +197,10 @@ class DataChannelConfigViewController: UITableViewController {
     // 開始時カメラ有効の入力値を configuration に渡します
     configuration.initialCameraEnabled =
       cameraEnabledOnConnectSegmentedControl.selectedSegmentIndex == 0
+
+    // 開始時マイク有効の入力値を configuration に渡します
+    configuration.initialMicrophoneEnabled =
+      microphoneEnabledOnConnectSegmentedControl.selectedSegmentIndex == 0
 
     return configuration
   }