+++ CHANGES.md
@@ -44,6 +44,10 @@
 - [UPDATE] 前面カメラ・背面カメラの切り替えボタンアイコンを置き換える
   - `camera` から `camera.rotate` に置き換える
   - @t-miya
+- [ADD] 接続メニューに `開始時マイク有効` 項目を追加する
+  - DataChannel、Simulcast、Spotlight、VideoChat が対象
+  - `無効` で接続した場合は音声ハードミュート相当の状態で開始する
+  - @t-miya
 - [ADD] 音声ハードミュート機能を追加する
   - DataChannel、Simulcast、Spotlight、VideoChat が対象
   - 録音 -> ソフトミュート -> ハードミュート -> 録音 と遷移するようにマイクミュートボタンの挙動を変更する
@@ -52,7 +56,7 @@
   - @t-miya
 - [ADD] 接続メニューに `接続時カメラ有効` 項目を追加する
   - DataChannel、Simulcast、Spotlight、VideoChat が対象
-  - `無効` で接続した場合はカメラハードミュート状態で開始する
+  - `無効` で接続した場合は映像ハードミュート相当の状態で開始する
   - @t-miya
 - [ADD] カメラハードミュート機能を追加する
   - DataChannel、Simulcast、Spotlight、VideoChat が対象